Move parser module to sexe-parser crate
