+++ sexe-parser/Cargo.toml
@@ -1,5 +1,8 @@
 [package]
-name = "sexe-parser"
+name = "sexe_parser"
 version = "0.1.0"
 
 [dependencies]
+sexe_expression = { version = "0.1.0", path = "../sexe-expression"}
+nom     = "^4.1"
+
